Migrating toward a single picker for ICD11 and MedDRA
