+++ packages/optimise-ui/src/components/treePicker/index.js
@@ -1,0 +1,4 @@
+import TreePickerHandle from './treePicker';
+
+export default TreePickerHandle;
+export const TreePicker = TreePickerHandle;
